+++ CHANGELOG.md
@@ -2,6 +2,7 @@ See `UPGRADE.md` for instructions on upgrading between versions.
 
 ## 0.49.0-dev
 
+* BREAKING: Removed `mogrify`/`imagemagick` -- use `sharp-cli`/`sharp` instead.
 * Add `?vsn=d` to all `fonts.pcss` URLs to fix fonts not caching.
 * Dynamic redirects.
 * Live preview: Changed syntax - see UPGRADE.md